+++ R/aliases.R
@@ -42,3 +42,10 @@ NEITHER <- function(x, y) !x && !y
 #' @rdname aliases
 #' @export
 pow <- `^`
+
+#' @rdname aliases
+#' @export
+XOR <- function(x, y) if (anyNA(x)) NA else if (x) !y else y
+# speed: anyNA > is.na > is.na || is.na
+
+