+++ NEWS.md
@@ -1,6 +1,13 @@
+# hutils 1.2.0
+* New functions:
+  - `RQ(p, yes, no)` short for `if (!requireNamespace("p", quietly = TRUE)) yes else no`.
+  - `isAttached` for conveniently determining whether a namespace is attached
+  - `ahull` for locating rectangles in a plot, as for automatically locating a text box.
+
 # hutils 1.1.1
 * Change of stable test:
     - The only changes were so that the test proceeds when a required package is present. The tests would never have succeeded in cases where they now skip. Consequently the major version number has not been incremented.
+    
 * Bug fix:
     - `find_pattern_in` respects `include.comments`
 * Enhancement: